refactor whisper/azure
