+++ mobile/examples/whisper/azure/android/app/build.gradle
@@ -5,11 +5,11 @@ plugins {
 }
 
 android {
-    namespace 'ai.onnxruntime.example.speechrecognition'
+    namespace 'ai.onnxruntime.example.whisperAzure'
     compileSdk 33
 
     defaultConfig {
-        applicationId "ai.onnxruntime.example.speechrecognition"
+        applicationId "ai.onnxruntime.example.whisperAzure"
         minSdk 24
         targetSdk 33
         versionCode 1
