+++ mobile/examples/whisper/azure/android/app/src/androidTest/java/ai/onnxruntime/example/whisperAzure/SpeechRecognitionInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package ai.onnxruntime.example.speechrecognition
+package ai.onnxruntime.example.whisperAzure
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
