+++ mobile/examples/whisper/azure/android/app/src/main/java/ai/onnxruntime/example/whisperAzure/AudioTensorSource.kt
@@ -1,4 +1,4 @@
-package ai.onnxruntime.example.speechrecognition
+package ai.onnxruntime.example.whisperAzure
 
 import ai.onnxruntime.OnnxJavaType
 import ai.onnxruntime.OnnxTensor
