+++ mobile/examples/whisper/azure/android/app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ai.onnxruntime.example.speechrecognition.MainActivity">
+    tools:context="ai.onnxruntime.example.whisperAzure.ai.onnxruntime.example.whisperAzure.MainActivity">
 
     <ImageView
         android:id="@+id/logo_image_view"
